Faster 3d building load
Created workaround in the Mapbox codebase that forces buildings to load using coroutines. There is a bug in the Mapbox codebase that prevents coroutines from working, even when an optimized tileset is defined, coroutines are toggled on, and a number of coroutines is specified. This fix hardcodes the max number of simultaneous buildings to load.

The result is we can easily pan the camera when using 3d buildings without much lag, resulting in a far-better user experience.